[Self-Review] Move `transform-api-error` to `common` and update project documentation
