+++ frontend/src/app/[locale]/(auth)/verify-code/page.tsx
@@ -12,7 +12,7 @@ import CodeInput from '@/design-system/components/code-input';
 import { Title } from '@/design-system/components/title';
 import { Button } from '@/design-system/ui/button';
 import { Separator } from '@/design-system/ui/separator';
-import { transformApiError } from '@/design-system/utils/transform-api-error';
+import { transformApiError } from '@/common/utils/transform-api-error';
 import { Link } from '@/i18n/navigation';
 import { useRouter } from '@/i18n/navigation';
