--- frontend/src/design-system/README.md
@@ -1,1 +1,0 @@
-# RuneBingo Design System
